# Quick and dirty performance timer