--- a/src/main.cpp
+++ b/src/main.cpp
@@ -54,6 +54,7 @@ extern "C" DLLEXPORT bool SKSEAPI SKSEPlugin_Load(const SKSE::LoadInterface* a_s
 	logger::info("FacegenFixes loaded");
 
 	SKSE::Init(a_skse);
+	SKSE::AllocTrampoline(14);
 
 	Hooks::Install();