Preprocessed FaceGen:
```
src\main.cpp(31): [info] FacegenFixes v1.0.1.0
src\main.cpp(54): [info] FacegenFixes loaded
src\SKSE\Trampoline.cpp(56): [info] Default Trampoline => 0B / 28B (00.00%)
src\FaceGenManager.cpp(56): [info] Installed hooks for face discoloration fix
src\SKSE\Trampoline.cpp(56): [info] Default Trampoline => 14B / 28B (50.00%)
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Tilma the Haggard' (   13bb6).
src\FaceGenManager.cpp(122): [info] Finished in 156450µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Vignar Gray-Mane' (   13bb5).
src\FaceGenManager.cpp(122): [info] Finished in 69335µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Aela the Huntress' (   1a696).
src\FaceGenManager.cpp(122): [info] Finished in 202µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Brill' (   1a6a2).
src\FaceGenManager.cpp(122): [info] Finished in 187µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Skjor' (   1a690).
src\FaceGenManager.cpp(122): [info] Finished in 247µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Prisoner' (       7).
src\FaceGenManager.cpp(122): [info] Finished in 1196µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Athis' (   1a6d5).
src\FaceGenManager.cpp(122): [info] Finished in 274µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Njada Stonearm' (   1a6d9).
src\FaceGenManager.cpp(122): [info] Finished in 264µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Torvar' (   1a6db).
src\FaceGenManager.cpp(122): [info] Finished in 244µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Ria' (   1a6d7).
src\FaceGenManager.cpp(122): [info] Finished in 214µs.
```

Runtime FaceGen:
```
src\main.cpp(31): [info] FacegenFixes v1.0.1.0
src\main.cpp(54): [info] FacegenFixes loaded
src\SKSE\Trampoline.cpp(56): [info] Default Trampoline => 0B / 28B (00.00%)
src\FaceGenManager.cpp(56): [info] Installed hooks for face discoloration fix
src\SKSE\Trampoline.cpp(56): [info] Default Trampoline => 14B / 28B (50.00%)
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Tilma the Haggard' (   13bb6).
src\FaceGenManager.cpp(122): [info] Finished in 180390µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Skjor' (   1a690).
src\FaceGenManager.cpp(122): [info] Finished in 66196µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Aela the Huntress' (   1a696).
src\FaceGenManager.cpp(122): [info] Finished in 52155µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Vignar Gray-Mane' (   13bb5).
src\FaceGenManager.cpp(122): [info] Finished in 57711µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Brill' (   1a6a2).
src\FaceGenManager.cpp(122): [info] Finished in 2440µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Prisoner' (       7).
src\FaceGenManager.cpp(122): [info] Finished in 1285µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Athis' (   1a6d5).
src\FaceGenManager.cpp(122): [info] Finished in 45946µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Njada Stonearm' (   1a6d9).
src\FaceGenManager.cpp(122): [info] Finished in 67359µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Torvar' (   1a6db).
src\FaceGenManager.cpp(122): [info] Finished in 32294µs.
src\FaceGenManager.cpp(116): [info] Getting head for NPC 'Ria' (   1a6d7).
src\FaceGenManager.cpp(122): [info] Finished in 131423µs.
```